# public/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Fichas Pacientes",
        "description": "Fichas Pacientes",
        "contact": {
            "email": "tss.labsi@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/enviar-json": {
            "post": {
                "tags": [
                    "JSON"
                ],
                "summary": "Registrar pacientes e suas visitas",
                "description": "",
                "operationId": "storeJson",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Internal error"
                    }
                }
            }
        },
        "/paciente": {
            "get": {
                "tags": [
                    "Paciente"
                ],
                "summary": "Pega lista de pacientes",
                "description": "Retorna lista de pacientes",
                "operationId": "pegaPacienteLista",
                "parameters": [
                    {
                        "name": "filter",
                        "in": "query",
                        "description": "filter",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PacienteResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/enviar-xml": {
            "post": {
                "tags": [
                    "XML"
                ],
                "summary": "Registrar pacientes e suas visitas",
                "description": "",
                "operationId": "storeXml",
                "requestBody": {
                    "content": {
                        "application/xml": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Internal error"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "JsonResource": {
                "title": "JsonResource",
                "description": "JsonResource",
                "required": [],
                "properties": {
                    "nome": {
                        "title": "nome",
                        "description": "nome do paciente",
                        "type": "string",
                        "example": "Tarcisio Souto"
                    },
                    "dtnasc": {
                        "title": "dtnasc",
                        "description": "data de nascimento",
                        "type": "string",
                        "example": "1993-02-26"
                    },
                    "sexo": {
                        "title": "sexo",
                        "description": "sexo",
                        "type": "string",
                        "example": "m"
                    },
                    "": {
                        "title": "consultas",
                        "description": "consultas",
                        "type": "array",
                        "items": {
                            "properties": {
                                "dtvisita": {
                                    "type": "string",
                                    "format": "date",
                                    "example": "2024-05-01"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
            "PacienteResource": {
                "title": "Paciente",
                "description": "Paciente",
                "required": [],
                "properties": {
                    "nome": {
                        "title": "nome",
                        "description": "nome",
                        "type": "string",
                        "example": "Tarcisio Souto"
                    },
                    "dtnasc": {
                        "title": "dtnasc",
                        "description": "data de nascimento",
                        "type": "string",
                        "format": "date",
                        "example": "1993-02-26"
                    },
                    "sexo": {
                        "title": "sexo",
                        "description": "sexo",
                        "type": "string",
                        "example": "m"
                    }
                },
                "type": "object"
            },
            "XmlResource": {
                "title": "XmlResource",
                "description": "XmlResource",
                "required": [],
                "properties": {
                    "nome": {
                        "title": "nome",
                        "description": "nome do paciente",
                        "type": "string",
                        "example": "Tarcisio Souto"
                    },
                    "dtnasc": {
                        "title": "dtnasc",
                        "description": "data de nascimento",
                        "type": "string",
                        "example": "1993-02-26"
                    },
                    "sexo": {
                        "title": "sexo",
                        "description": "sexo",
                        "type": "string",
                        "example": "m"
                    },
                    "": {
                        "title": "consultas",
                        "description": "consultas",
                        "type": "array",
                        "items": {
                            "properties": {
                                "dtvisita": {
                                    "type": "string",
                                    "format": "date",
                                    "example": "2024-05-01"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            }
        }
    }
}